Systemsimulation/Projekt_Sandra_Stroiczek/invasive_plants_simulation/requirements.txt

numpy
matplotlib
Pillow
imageio
